+++ kipper/core/src/compiler/ast/nodes/declarations/function-declaration/function-declaration-type-semantics.ts
@@ -2,7 +2,8 @@
  * Type semantics for AST Node {@link FunctionDeclaration}.
  * @since 0.10.0
  */
-import type { ProcessedType } from "../../../../semantics";
+import type { BuiltInTypeFunc } from "../../../../semantics";
+import { ProcessedType } from "../../../../semantics";
 import type { DeclarationTypeSemantics } from "../declaration-type-semantics";
 
 /**
@@ -11,8 +12,8 @@ import type { DeclarationTypeSemantics } from "../declaration-type-semantics";
  */
 export interface FunctionDeclarationTypeSemantics extends DeclarationTypeSemantics {
 	/**
-	 * The {@link KipperType return type} of the function.
-	 * @since 0.10.0
+	 * The type of the declaration. This is always some variation of {@link BuiltInTypeFunc i.e. `Func<P..., T>`}.
+	 * @since 0.12.0
 	 */
-	returnType: ProcessedType;
+	type: BuiltInTypeFunc;
 }