+++ CHANGELOG.md
@@ -18,9 +18,9 @@ To use development versions of Kipper download the
 
 ### Added
 
-- Added semantic checking and code generation for classes, class methods, class properties and class constructors.
+- Support for classes, class methods, class properties and class constructors.
   ([#665](https://github.com/Kipper-Lang/Kipper/issues/665))
-- Added semantic checking and code generation for object literals and object properties.
+- Support for object literals and object properties.
   ([#526](https://github.com/Kipper-Lang/Kipper/issues/526))
 - Implemented internal representation for custom types such as objects, interfaces and classes. This change means that
   the entire core type system has been reworked and adjusted to also support custom types as well as complex types
@@ -28,12 +28,13 @@ To use development versions of Kipper download the
   implementation of all custom types in the future. ([#524](https://github.com/Kipper-Lang/Kipper/issues/524))
 - Implemented the generic `Array<T>` type and single-type array initializers.
   ([#499](https://github.com/Kipper-Lang/Kipper/issues/499))
+- Support for index-based array assignments. ([#669](https://github.com/Kipper-Lang/Kipper/issues/669))
 - Implemented the generic `Func<T..., R>` type and function type initializers.
   ([#584](https://github.com/Kipper-Lang/Kipper/issues/584))
 - Implemented internal generic spread argument `T...`, which allows multiple arguments to be passed to a single
   parameter inside of a generic type specifier.
 - Implemented constant `NaN`, which represents the `NaN` value in JavaScript (Not a Number).
-	([#671](https://github.com/Kipper-Lang/Kipper/issues/671))
+  ([#671](https://github.com/Kipper-Lang/Kipper/issues/671))
 - New module:
   - `semantics/runtime-built-ins`, which contains runtime built-in functions, variables and types.
   - `semantics/runtime-internals`, which contains the runtime internal functions.
@@ -61,7 +62,7 @@ To use development versions of Kipper download the
     an error indicating an invalid logic that should be fixed.
   - `MismatchingArgCountBetweenFuncTypesError`, which is thrown when the amount of arguments in a function type does not
     match the number of arguments in the function type it is compared to.
-- New interfaces:
+- New interfaces and types:
   - `InterfaceDeclarationSemantics`, which represents the semantics of an interface declaration.
   - `InterfaceDeclarationTypeSemantics`, which represents the type semantics of an interface declaration.
   - `ClassMethodDeclarationSemantics`, which represents the semantics of a class method declaration.
@@ -76,8 +77,15 @@ To use development versions of Kipper download the
   - `TypeDeclarationSemantics`, which represents the semantics of a type declaration.
   - `TypeDeclarationTyp`KipperTypeChecker.validArrayExpression`eSemantics`, which represents the type semantics of a type declaration.
   - `CompilableType`, which represents a type that can be compiled.
+  - `BuiltInReference`, which replaces the now removed type `Reference` in the `KipperProgramContext` for reference
+    tracking of built-in types.
 - New functions:
   - `KipperTypeChecker.validArrayExpression`, which ensures that an array expression is valid.
+- New Properties:
+  - `BuiltInFunction.funcType`, which returns a function type for the built-in function.
+  - `FunctionDeclarationTypeSemantics.type`, which returns the type of the function declaration i.e. the function type.
+  - `LambdaPrimaryExpressionTypeSemantics.type`, which returns the type of the lambda primary expression i.e. the
+    function type.
 
 ### Changed
 
@@ -88,15 +96,21 @@ To use development versions of Kipper download the
   - `BuiltInVariable`, which represents a built-in variable.
 - Renamed:
   - Module `analysis` to `semantics`.
+  - Module `compiler/.../expressions/arithmetic` to `arithmetic-expression`.
   - Class `UncheckedType` to `RawType`.
   - Class `CheckedType` to `ProcessedType`.
   - Class `UndefinedCustomType` to `UndefinedType`.
 
 ### Fixed
 
+- All functions and lambdas simply resolving to `Func` instead of the appropriate filled-up `Func<T..., R>` type. This
+  now enables proper type checking for function references.
+
 ### Deprecated
 
 ### Removed
+
+- Type `Reference` as it is no longer needed and has been replaced by `KipperReferenceable`.
 
 </details>